Move ClassDefinition
