+++ src/ClassDefinition/ClassDefinition.php
@@ -2,7 +2,12 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilCompilationSource;
+namespace webignition\BasilCompilationSource\ClassDefinition;
+
+use webignition\BasilCompilationSource\Metadata;
+use webignition\BasilCompilationSource\MetadataInterface;
+use webignition\BasilCompilationSource\MethodDefinitionInterface;
+use webignition\BasilCompilationSource\SourceInterface;
 
 class ClassDefinition implements ClassDefinitionInterface
 {
