+++ src/ClassDefinition/ClassDefinitionInterface.php
@@ -2,7 +2,10 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilCompilationSource;
+namespace webignition\BasilCompilationSource\ClassDefinition;
+
+use webignition\BasilCompilationSource\MethodDefinitionInterface;
+use webignition\BasilCompilationSource\SourceInterface;
 
 interface ClassDefinitionInterface extends SourceInterface
 {
