+++ tests/Unit/ClassDefinition/ClassDefinitionTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilCompilationSource\Tests\Unit;
+namespace webignition\BasilCompilationSource\Tests\Unit\ClassDefinition;
 
-use webignition\BasilCompilationSource\ClassDefinition;
-use webignition\BasilCompilationSource\ClassDefinitionInterface;
+use webignition\BasilCompilationSource\ClassDefinition\ClassDefinition;
+use webignition\BasilCompilationSource\ClassDefinition\ClassDefinitionInterface;
 use webignition\BasilCompilationSource\ClassDependency;
 use webignition\BasilCompilationSource\ClassDependencyCollection;
 use webignition\BasilCompilationSource\Comment;
